choosing-a-code-editor: Expanded and clarified descriptions
For VSCode, the Haxe extension is now mentioned as a requirement. There were questions on the forums about whether Haxe and Lime extensions conflicted, so best to be clearer about that.

Some of the descriptions seem to have been written before VSCode was added to the list, and they still seemed to suggest that FlashDevelop was the best option, even though we we specifically recommended VSCode at the top.